added explicit typing for ci tests

+++ graphrag/api/query.py
@@ -301,9 +301,7 @@ async def multi_global_search(
             }
         community_reports_df["community"] += max_vals["community_reports"] + 1
         community_reports_df["human_readable_id"] += max_vals["community_reports"] + 1
-        max_vals["community_reports"] = (
-            community_reports_df["community"].astype(int).max()
-        )
+        max_vals["community_reports"] = int(community_reports_df["community"].max())
         community_reports_dfs.append(community_reports_df)
 
         # Prepare each index's communities dataframe for merging
@@ -320,7 +318,7 @@ async def multi_global_search(
             lambda x: x if x == -1 else x + max_vals["community"] + 1
         )
         communities_df["human_readable_id"] += max_vals["community"] + 1
-        max_vals["community"] = communities_df["community"].astype(int).max()
+        max_vals["community"] = int(communities_df["community"].max())
         communities_dfs.append(communities_df)
 
         # Prepare each index's entities dataframe for merging
@@ -337,7 +335,7 @@ async def multi_global_search(
         entities_df["text_unit_ids"] = entities_df["text_unit_ids"].apply(
             lambda x, index_name=index_name: [i + f"-{index_name}" for i in x]
         )
-        max_vals["entities"] = entities_df["human_readable_id"].astype(int).max()
+        max_vals["entities"] = int(entities_df["human_readable_id"].max())
         entities_dfs.append(entities_df)
 
     # Merge the dataframes
@@ -616,7 +614,7 @@ async def multi_local_search(
         nodes_df["id"] = nodes_df["id"].apply(
             lambda x, index_name=index_name: x + f"-{index_name}"
         )
-        max_vals["nodes"] = int(nodes_df["human_readable_id"].astype(int).max())
+        max_vals["nodes"] = int(nodes_df["human_readable_id"].max())
         nodes_dfs.append(nodes_df)
 
         # Prepare each index's community reports dataframe for merging
@@ -631,9 +629,7 @@ async def multi_local_search(
             }
         community_reports_df["community"] += max_vals["community_reports"] + 1
         community_reports_df["human_readable_id"] += max_vals["community_reports"] + 1
-        max_vals["community_reports"] = (
-            community_reports_df["community"].astype(int).max()
-        )
+        max_vals["community_reports"] = int(community_reports_df["community"].max())
         community_reports_dfs.append(community_reports_df)
 
         # Prepare each index's entities dataframe for merging
@@ -653,7 +649,7 @@ async def multi_local_search(
         entities_df["text_unit_ids"] = entities_df["text_unit_ids"].apply(
             lambda x, index_name=index_name: [i + f"-{index_name}" for i in x]
         )
-        max_vals["entities"] = entities_df["human_readable_id"].astype(int).max()
+        max_vals["entities"] = int(entities_df["human_readable_id"].max())
         entities_dfs.append(entities_df)
 
         # Prepare each index's relationships dataframe for merging
@@ -679,9 +675,7 @@ async def multi_local_search(
         relationships_df["text_unit_ids"] = relationships_df["text_unit_ids"].apply(
             lambda x, index_name=index_name: [i + f"-{index_name}" for i in x]
         )
-        max_vals["relationships"] = (
-            relationships_df["human_readable_id"].astype(int).max()
-        )
+        max_vals["relationships"] = int(relationships_df["human_readable_id"].max())
         relationships_dfs.append(relationships_df)
 
         # Prepare each index's text units dataframe for merging
